AI Chat: Move everything from brave/browser/ai_chat back to brave/components/ai_chat
Didn't split to content, etc for now
