+++ browser/sources.gni
@@ -110,8 +110,6 @@ brave_chrome_browser_deps = [
   "//brave/app:brave_generated_resources_grit",
   "//brave/browser:browser_process",
   "//brave/browser:sparkle_buildflags",
-  "//brave/browser/ai_chat",
-  "//brave/browser/ai_chat:mojom",
   "//brave/browser/brave_federated",
   "//brave/browser/brave_wallet",
   "//brave/browser/brave_wallet:brave_wallet_delegate",
@@ -131,6 +129,7 @@ brave_chrome_browser_deps = [
   "//brave/browser/ui/tabs",
   "//brave/common",
   "//brave/components/ai_chat",
+  "//brave/components/ai_chat:mojom",
   "//brave/components/brave_ads/browser",
   "//brave/components/brave_ads/common",
   "//brave/components/brave_component_updater/browser",
