+++ browser/ui/webui/ai_chat/ai_chat_ui.cc
@@ -7,11 +7,11 @@
 
 #include <utility>
 
-#include "brave/browser/ai_chat/constants.h"
-#include "brave/browser/ai_chat/pref_names.h"
 #include "brave/browser/ui/side_panel/ai_chat/ai_chat_side_panel_utils.h"
 #include "brave/browser/ui/webui/ai_chat/ai_chat_ui_page_handler.h"
 #include "brave/browser/ui/webui/brave_webui_source.h"
+#include "brave/components/ai_chat/constants.h"
+#include "brave/components/ai_chat/pref_names.h"
 #include "brave/components/ai_chat_ui/grit/ai_chat_ui_generated_map.h"
 #include "brave/components/constants/webui_url_constants.h"
 #include "brave/components/l10n/common/localization_util.h"
