+++ components/ai_chat/DEPS
@@ -1,0 +1,5 @@
+include_rules = [
+  "+content/public/browser",
+  "+services/network/public",
+  "+ui",
+]
