+++ components/ai_chat/ai_chat_api.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_BROWSER_AI_CHAT_AI_CHAT_API_H_
-#define BRAVE_BROWSER_AI_CHAT_AI_CHAT_API_H_
+#ifndef BRAVE_COMPONENTS_AI_CHAT_AI_CHAT_API_H_
+#define BRAVE_COMPONENTS_AI_CHAT_AI_CHAT_API_H_
 
 #include <string>
 
@@ -39,4 +39,4 @@ class AIChatAPI {
   base::WeakPtrFactory<AIChatAPI> weak_ptr_factory_{this};
 };
 
-#endif  // BRAVE_BROWSER_AI_CHAT_AI_CHAT_API_H_
+#endif  // BRAVE_COMPONENTS_AI_CHAT_AI_CHAT_API_H_
