+++ components/ai_chat/pref_names.cc
@@ -3,18 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_BROWSER_AI_CHAT_PREF_NAMES_H_
-#define BRAVE_BROWSER_AI_CHAT_PREF_NAMES_H_
+#include "brave/components/ai_chat/pref_names.h"
 
-class PrefRegistrySimple;
+#include "components/prefs/pref_registry_simple.h"
 
 namespace ai_chat::prefs {
 
-constexpr char kBraveChatHasSeenDisclaimer[] =
-    "brave.ai_chat.has_seen_disclaimer";
-
-void RegisterProfilePrefs(PrefRegistrySimple* registry);
+void RegisterProfilePrefs(PrefRegistrySimple* registry) {
+  registry->RegisterBooleanPref(kBraveChatHasSeenDisclaimer, false);
+}
 
 }  // namespace ai_chat::prefs
-
-#endif  // BRAVE_BROWSER_AI_CHAT_PREF_NAMES_H_
