+++ components/ai_chat/pref_names.h
@@ -3,14 +3,18 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/browser/ai_chat/pref_names.h"
+#ifndef BRAVE_COMPONENTS_AI_CHAT_PREF_NAMES_H_
+#define BRAVE_COMPONENTS_AI_CHAT_PREF_NAMES_H_
 
-#include "components/prefs/pref_registry_simple.h"
+class PrefRegistrySimple;
 
 namespace ai_chat::prefs {
 
-void RegisterProfilePrefs(PrefRegistrySimple* registry) {
-  registry->RegisterBooleanPref(kBraveChatHasSeenDisclaimer, false);
-}
+constexpr char kBraveChatHasSeenDisclaimer[] =
+    "brave.ai_chat.has_seen_disclaimer";
+
+void RegisterProfilePrefs(PrefRegistrySimple* registry);
 
 }  // namespace ai_chat::prefs
+
+#endif  // BRAVE_COMPONENTS_AI_CHAT_PREF_NAMES_H_
