+++ components/ai_chat_ui/BUILD.gn
@@ -11,7 +11,7 @@ transpile_web_ui("ai_chat_ui") {
         "chat_ui",
         rebase_path("chat_ui.tsx"),
       ] ]
-  deps = [ "//brave/browser/ai_chat:mojom_js" ]
+  deps = [ "//brave/components/ai_chat:mojom_js" ]
 }
 
 pack_web_resources("generated_resources") {
